Recipes rework. Adapted from: L2jUnity free files.

+++ L2J_Mobius_4.0_GrandCrusade/java/com/l2jmobius/gameserver/model/L2Seed.java
@@ -32,8 +32,8 @@ public final class L2Seed
 	private final boolean _isAlternative;
 	private final int _limitSeeds;
 	private final int _limitCrops;
-	private final int _seedReferencePrice;
-	private final int _cropReferencePrice;
+	private final long _seedReferencePrice;
+	private final long _cropReferencePrice;
 	
 	public L2Seed(StatsSet set)
 	{
@@ -99,27 +99,27 @@ public final class L2Seed
 		return _limitCrops * Config.RATE_DROP_MANOR;
 	}
 	
-	public final int getSeedReferencePrice()
+	public final long getSeedReferencePrice()
 	{
 		return _seedReferencePrice;
 	}
 	
-	public final int getSeedMaxPrice()
+	public final long getSeedMaxPrice()
 	{
 		return _seedReferencePrice * 10;
 	}
 	
-	public final int getSeedMinPrice()
+	public final long getSeedMinPrice()
 	{
 		return (int) (_seedReferencePrice * 0.6);
 	}
 	
-	public final int getCropReferencePrice()
+	public final long getCropReferencePrice()
 	{
 		return _cropReferencePrice;
 	}
 	
-	public final int getCropMaxPrice()
+	public final long getCropMaxPrice()
 	{
 		return _cropReferencePrice * 10;
 	}
