+++ L2J_Mobius_4.0_GrandCrusade/java/com/l2jmobius/gameserver/network/serverpackets/SellList.java
@@ -76,7 +76,7 @@ public class SellList implements IClientOutgoingPacket
 		
 		for (L2ItemInstance item : _sellList)
 		{
-			int price = item.getItem().getReferencePrice() / 2;
+			long price = item.getItem().getReferencePrice() / 2;
 			if (_merchant != null)
 			{
 				price -= (price * _merchant.getTotalTaxRate(TaxType.SELL));
